% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castarterpresidents.R
\docType{package}
\name{castarterpresidents}
\alias{castarterpresidents}
\alias{castarterpresidents-package}
\title{Datasets of press releases issued by presidents}
\description{
Datasets of presidential press releases compatible with the \code{castarter} package
}
\details{
castarterpresidents
Datasets of presidential press releases compatible with the \code{castarter} package

Use \code{data(package='castarterpresidents')$results[, 3]} tosee a list of availabledata sets in this data package
}
\seealso{
\link{kremlin_en}
}
