% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castarterpresidents.R
\docType{data}
\name{kremlin_en}
\alias{kremlin_en}
\title{Press releases issued by Russia's president}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{doc_id}{}
\item{text}{}
\item{website}{}
\item{id}{}
\item{date}{}
\item{title}{}
\item{language}{}
\item{link}{}
}}
\source{
The data comes from kremlin.ru.
}
\description{
Press releases from the Kremlin's official website
}
\seealso{
\link{castarterpresidents}
}
